@@ -18,77 +18,79 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^7.2.11",
-    "@angular/common": "^7.2.11",
-    "@angular/compiler": "^7.2.11",
-    "@angular/core": "^7.2.11",
-    "@angular/forms": "^7.2.11",
-    "@angular/http": "^7.2.11",
-    "@angular/platform-browser": "^7.2.11",
-    "@angular/platform-browser-dynamic": "^7.2.11",
-    "@angular/router": "^7.2.11",
-    "@fortawesome/angular-fontawesome": "^0.3.0",
-    "@fortawesome/fontawesome-svg-core": "^1.2.17",
-    "@fortawesome/free-regular-svg-icons": "^5.8.1",
-    "@fortawesome/free-solid-svg-icons": "^5.8.1",
-    "@google-cloud/storage": "^2.5.0",
-    "@ionic-native/google-plus": "^5.3.0",
+    "@angular/animations": "^8.2.7",
+    "@angular/common": "^8.2.7",
+    "@angular/compiler": "^8.2.7",
+    "@angular/core": "^8.2.7",
+    "@angular/forms": "^8.2.7",
+    "@angular/platform-browser": "^8.2.7",
+    "@angular/platform-browser-dynamic": "^8.2.7",
+    "@angular/pwa": "^0.803.9",
+    "@angular/router": "^8.2.7",
+    "@angular/service-worker": "^8.2.7",
+    "@fortawesome/angular-fontawesome": "^0.5.0",
+    "@fortawesome/fontawesome-svg-core": "^1.2.24",
+    "@fortawesome/free-regular-svg-icons": "^5.11.1",
+    "@fortawesome/free-solid-svg-icons": "^5.11.1",
+    "@google-cloud/storage": "^3.3.0",
+    "@ionic-native/google-plus": "^5.14.0",
     "@types/electron": "^1.6.10",
     "@types/hashids": "^1.0.30",
-    "angular-oauth2-oidc": "^5.0.2",
-    "axios": "^0.18.0",
-    "blueimp-canvas-to-blob": "^3.14.0",
+    "angular-oauth2-oidc": "^8.0.4",
+    "axios": "^0.19.0",
+    "blueimp-canvas-to-blob": "^3.16.0",
     "bootstrap": "^4.3.1",
-    "core-js": "^3.0.0",
+    "core-js": "^3.2.1",
     "cors": "^2.8.5",
-    "custom-electron-titlebar": "^3.0.8",
-    "electron-debug": "^2.1.0",
-    "firebase": "^5.11.1",
-    "firebase-admin": "^7.3.0",
-    "firebase-functions": "^2.3.1",
+    "custom-electron-titlebar": "^3.1.0",
+    "electron-debug": "^3.0.1",
+    "firebase": "^6.6.2",
+    "firebase-admin": "^8.6.0",
+    "firebase-functions": "^3.2.0",
     "font-awesome": "^4.7.0",
     "fs": "0.0.1-security",
-    "hashids": "^1.2.2",
     "jquery": "^3.4.1",
     "ng2-img-max": "^2.1.18",
     "ng2-pdf-viewer": "^5.3.4",
     "ng2-pdfjs-viewer": "^5.0.1",
     "ngx-electron": "^2.1.1",
     "ngx-smart-modal": "^7.1.1",
-    "nodemon": "^1.18.10",
-    "rxjs": "~6.4.0",
-    "ts-md5": "^1.2.4",
+    "nodemon": "^1.19.2",
+    "pdfkit": "^0.10.0",
+    "postmark": "^2.3.0",
+    "rxjs": "~6.5.3",
+    "ts-md5": "^1.2.6",
     "underscore": "latest",
-    "zone.js": "~0.9.0"
+    "zone.js": "~0.9.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.13.9",
-    "@angular/cli": "~7.3.7",
-    "@angular/compiler-cli": "^7.2.11",
-    "@angular/fire": "^5.1.3",
-    "@angular/language-service": "^7.2.11",
-    "@types/jasmine": "~3.3.12",
+    "@angular-devkit/build-angular": "~0.803.5",
+    "@angular/cli": "~8.3.5",
+    "@angular/compiler-cli": "^8.2.7",
+    "@angular/fire": "^5.2.1",
+    "@angular/language-service": "^8.2.7",
+    "@types/jasmine": "^3.4.0",
     "@types/jasminewd2": "~2.0.6",
-    "@types/node": "~11.12.1",
-    "angular-cli-ghpages": "^0.5.3",
-    "codelyzer": "~5.0.0",
+    "@types/node": "~12.7.5",
+    "angular-cli-ghpages": "^0.6.0",
+    "codelyzer": "^5.0.1",
     "devtron": "^1.4.0",
-    "electron": "^5.0.1",
-    "electron-packager": "^13.1.1",
+    "electron": "^6.0.10",
+    "electron-packager": "^14.0.6",
     "electron-squirrel-startup": "^1.0.0",
-    "electron-winstaller": "^3.0.4",
+    "electron-winstaller": "^4.0.0",
     "firebase-functions-test": "^0.1.6",
-    "jasmine-core": "~3.3.0",
+    "jasmine-core": "~3.5.0",
     "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~4.0.1",
-    "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "~2.0.5",
+    "karma": "~4.3.0",
+    "karma-chrome-launcher": "~3.1.0",
+    "karma-coverage-istanbul-reporter": "^2.1.0",
     "karma-jasmine": "~2.0.1",
     "karma-jasmine-html-reporter": "^1.4.0",
     "node-sass": "^4.12.0",
     "protractor": "~6.0.0",
-    "ts-node": "~8.0.3",
-    "tslint": "^5.14.0",
-    "typescript": "^3.1.6"
+    "ts-node": "~8.4.1",
+    "tslint": "^5.20.0",
+    "typescript": "^3.5.3"
   }
 }
