@@ -1,7 +1,7 @@
 import {Component, OnInit, ViewChild} from '@angular/core';
 import {SnackbarComponent} from './tools/snackbar/snackbar.component';
 import {ToolsServices} from './services/tools-services.service';
-import {SHOW_WINDOWS_BAR} from '../environments/environment';
+import {SHOW_WINDOWS_BAR} from '../db/dbConfig';
 
 @Component({
   selector: 'app-root',
@@ -9,13 +9,12 @@ import {SHOW_WINDOWS_BAR} from '../environments/environment';
   styleUrls: ['./app.component.css']
 })
 export class AppComponent implements OnInit {
-  @ViewChild(SnackbarComponent)
+  @ViewChild(SnackbarComponent, {static: false})
   public snack: SnackbarComponent;
   public showBar = SHOW_WINDOWS_BAR;
   const;
 
   // customTitlebar = require('custom-electron-titlebar');
-
   /**
    * Esta clase se encarga de iniciar toods los sevicios apenas alguien entre a la pag.
    * Contiene el snakBar
