@@ -2,8 +2,9 @@ import {Component, OnInit} from '@angular/core';
 import {DbManagerService} from '../../../services/db-manager.service';
 import {ToolsServices} from '../../../services/tools-services.service';
 import {AuthService} from '../../../services/routes/auth.service';
-import {CurrentStorageService} from '../../../services/current-storage.service';
+import {BeforeAppInitService} from '../../../services/before-app-init.service';
 import {Router} from '@angular/router';
+import {CurrentDataService} from '../../../current-data.service';
 
 
 @Component({
@@ -14,12 +15,22 @@ import {Router} from '@angular/router';
 export class TopBarComponent implements OnInit {
   typeName = 'Piezas';
   iconStyle = ToolsServices.iconStyle;
+  scrumBreads: string[] = [];
+  selectedTab: any;
+  tabsData = CurrentDataService.topTabs;
 
-  constructor(public db: DbManagerService, public tools: ToolsServices,
-              public current: CurrentStorageService, private auth: AuthService, private router: Router) {
+  constructor(public db: DbManagerService, public tools: ToolsServices, private global: CurrentDataService,
+              public current: BeforeAppInitService, private auth: AuthService, private router: Router) {
   }
 
   ngOnInit() {
+    this.scrumBreads = CurrentDataService.currentRouteInfo.scrumBreads;
+    this.selectedTab = this.global.currentSelectedTab;
+    this.global.tobTabsEmitter.subscribe(value => {
+      this.tabsData = value;
+      this.scrumBreads = CurrentDataService.currentRouteInfo.scrumBreads;
+      this.selectedTab = this.global.currentSelectedTab;
+    });
 
   }
 
@@ -35,5 +46,9 @@ export class TopBarComponent implements OnInit {
     document.body.style.setProperty(`--main-color`, c.color);
     // this.current.whenChange.emit(c);
     this.current.productSelected = c;
+  }
+
+  toggleMenu() {
+    CurrentDataService.toggleMenu = !CurrentDataService.toggleMenu;
   }
 }
