@@ -2,34 +2,28 @@ import {Component, OnInit} from '@angular/core';
 import {fadeAnimation} from '../../animations/Animation';
 import {ToolsServices} from '../../services/tools-services.service';
 import {Router} from '@angular/router';
+import {CurrentDataService} from '../../current-data.service';
 
 @Component({
   selector: 'app-main',
   templateUrl: './main.component.html',
-  styleUrls: ['./main.component.css'],
+  styleUrls: ['./main.component.scss'],
   animations: [fadeAnimation]
 })
 
 export class MainComponent implements OnInit {
-  selectedItemTab: any = undefined;
+  first = true;
 
   constructor(public tools: ToolsServices,
+              public globalData: CurrentDataService,
               private router: Router) {
+
   }
 
   ngOnInit(): void {
-    this.selectedItemTab = this.tools.currentSelectedTab;
-  }
-
-  selectTab(tab: any) {
-    this.tools.currentSelectedTab = tab;
-    this.router.navigateByUrl(tab.path);
-  }
-
-  selectDtab(dynamicTab: any) {
-    if (dynamicTab !== this.tools.currentSelectedTab) {
-      this.tools.currentSelectedTab = dynamicTab;
-      this.router.navigate([dynamicTab.path], {queryParams: dynamicTab.queryParams});
+    if (this.first) {
+      this.globalData.loadTabDataByUrl(this.router.url);
+      this.first = false;
     }
   }
 }
