@@ -1,12 +1,12 @@
 import { TestBed } from '@angular/core/testing';
 
-import { CurrentStorageService } from './current-storage.service';
+import { BeforeAppInitService } from './before-app-init.service';
 
 describe('CurrentStorageService', () => {
   beforeEach(() => TestBed.configureTestingModule({}));
 
   it('should be created', () => {
-    const service: CurrentStorageService = TestBed.get(CurrentStorageService);
+    const service: BeforeAppInitService = TestBed.get(BeforeAppInitService);
     expect(service).toBeTruthy();
   });
 });
